+++ src/Arkivverket.Arkade.Core/Testing/ILocation.cs
@@ -5,7 +5,7 @@ namespace Arkivverket.Arkade.Core.Testing
     public interface ILocation
     {
         string FileName { get; }
-        IEnumerable<int> ErrorLocations { get; }
+        IEnumerable<long> ErrorLocations { get; }
         string ToString();
     }
 }
