+++ src/Arkivverket.Arkade.Core/Testing/Noark5/N5_17_NumberOfEachJournalPostType.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using System.Linq;
 using Arkivverket.Arkade.Core.Base;
 using Arkivverket.Arkade.Core.Base.Noark5;
 using Arkivverket.Arkade.Core.Resources;
@@ -96,7 +95,7 @@ namespace Arkivverket.Arkade.Core.Testing.Noark5
             if (eventArgs.Path.Matches("journalposttype", "registrering") && _journalPostAttributeIsFound)
             {
                 string journalPostType = eventArgs.Value;
-                int xmlLineNumber = eventArgs.LineNumber;
+                long xmlLineNumber = eventArgs.LineNumber;
                 if (_journalPostTypesPerArchivePart.ContainsKey(_currentArchivePart))
                 {
                     if (_journalPostTypesPerArchivePart[_currentArchivePart].ContainsKey(journalPostType))
@@ -136,12 +135,12 @@ namespace Arkivverket.Arkade.Core.Testing.Noark5
         private class JournalPostType
         {
             public int Count { get; set; }
-            public List<int> Locations { get; }
+            public List<long> Locations { get; }
 
-            public JournalPostType(int xmlLineNumber)
+            public JournalPostType(long xmlLineNumber)
             {
                 Count = 1;
-                Locations = new List<int> { xmlLineNumber };
+                Locations = new List<long> { xmlLineNumber };
             }
         }
     }
