+++ src/Arkivverket.Arkade.Core/Testing/Noark5/N5_47_SystemIdUniqueControl.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
 using Arkivverket.Arkade.Core.Base;
 using Arkivverket.Arkade.Core.Base.Noark5;
 using Arkivverket.Arkade.Core.Resources;
@@ -43,7 +42,7 @@ namespace Arkivverket.Arkade.Core.Testing.Noark5
             if (eventArgs.Path.Matches("systemID"))
             {
                 string systemId = eventArgs.Value;
-                int xmlLineNumber = eventArgs.LineNumber;
+                long xmlLineNumber = eventArgs.LineNumber;
 
                 if (_systemIdInstances.ContainsKey(systemId))
                 {
@@ -70,12 +69,12 @@ namespace Arkivverket.Arkade.Core.Testing.Noark5
         private class Instance
         {
             public int Count { get; set; }
-            public List<int> Locations { get; }
+            public List<long> Locations { get; }
 
-            public Instance(int xmlLineNumber)
+            public Instance(long xmlLineNumber)
             {
                 Count = 1;
-                Locations = new List<int>{xmlLineNumber};
+                Locations = new List<long> {xmlLineNumber};
             }
         }
     }
